Networking is finally unfucked
